+++ docs/history.rst
@@ -7,6 +7,7 @@ Changelog
 * buildout now used for testing [Russell]
 * removed dependency on xpath, adds 2.6 compatability [Russell]
 * changed name of https proxy urllib2 egg to match names recognised by eggproxy [Russell]
+* replaced custom parser with BeautifulSoup [Russell]
 
 0.1 - 2009-10-27
 ----------------